Renamespace integration tests
