+++ tests/Integration/User/CreateTokenTest.php
@@ -2,9 +2,10 @@
 
 declare(strict_types=1);
 
-namespace App\Tests\Integration;
+namespace App\Tests\Integration\User;
 
 use App\Tests\Application\User\AbstractCreateTokenTest;
+use App\Tests\Integration\GetClientAdapterTrait;
 
 class CreateTokenTest extends AbstractCreateTokenTest
 {
