+++ tests/Integration/User/GetDefaultApiKeyTest.php
@@ -2,9 +2,10 @@
 
 declare(strict_types=1);
 
-namespace App\Tests\Integration;
+namespace App\Tests\Integration\User;
 
 use App\Tests\Application\User\AbstractGetDefaultApiKeyTest;
+use App\Tests\Integration\GetClientAdapterTrait;
 
 class GetDefaultApiKeyTest extends AbstractGetDefaultApiKeyTest
 {
