+++ tests/Integration/User/RefreshTokenTest.php
@@ -2,9 +2,10 @@
 
 declare(strict_types=1);
 
-namespace App\Tests\Integration;
+namespace App\Tests\Integration\User;
 
 use App\Tests\Application\User\AbstractRefreshTokenTest;
+use App\Tests\Integration\GetClientAdapterTrait;
 
 class RefreshTokenTest extends AbstractRefreshTokenTest
 {
