+++ tests/Integration/User/VerifyTokenTest.php
@@ -2,9 +2,10 @@
 
 declare(strict_types=1);
 
-namespace App\Tests\Integration;
+namespace App\Tests\Integration\User;
 
 use App\Tests\Application\User\AbstractVerifyTokenTest;
+use App\Tests\Integration\GetClientAdapterTrait;
 
 class VerifyTokenTest extends AbstractVerifyTokenTest
 {
